move properties panel under element for easier tab categorization
