+++ packages/editor/src/components/EditorContainer.tsx
@@ -65,11 +65,11 @@ import SaveNewSceneDialog from './dialogs/SaveNewSceneDialog'
 import SaveSceneDialog from './dialogs/SaveSceneDialog'
 import { DndWrapper } from './dnd/DndWrapper'
 import DragLayer from './dnd/DragLayer'
+import { PropertiesPanelTab } from './element/PropertiesPanel'
 import { GraphPanelTab } from './graph/GraphPanel'
 import { HierarchyPanelTab } from './hierarchy/HierarchyPanel'
 import { MaterialLibraryPanelTab } from './materials/MaterialLibraryPanel'
 import { ViewportPanelTab } from './panels/ViewportPanel'
-import { PropertiesPanelTab } from './properties/PropertiesPanel'
 import * as styles from './styles.module.scss'
 import ToolBar from './toolbar/ToolBar'
